requirements.txt
flask
flasgger
mysql-connector-python
gunicorn
Flask-SQLAlchemy
Flask-Migrate
PyJWT
cryptography
